+++ src/components/TaskListColumn.vue
@@ -23,7 +23,7 @@ interface Props {
   cssClass?: string
 }
 
-const props = withDefaults(defineProps<Props>(), {
+withDefaults(defineProps<Props>(), {
   align: 'left',
 })